+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 Not released
 ------------
 
+* Add accessors for the internal state of the LDA.
+* Introduce the `scalib.ScalibError` exception and remove `scalib.metrics.SnrError` (**Breaking change**).
 * Allow all computations to be interrupted with Ctrl-C.
 * Fix deadlock when there is an error in large SNR computations (i.e., when n_vars*n_samples*n_traces > 2**33).
 * Allow LDA to behave like simple pooled gaussian templates (#22)